Add new Centros page
Update index. Closes #42
